+++ kgcnn/mol/convert.py
@@ -11,13 +11,13 @@ module_logger = logging.getLogger(__name__)
 module_logger.setLevel(logging.INFO)
 
 try:
-    from kgcnn.mol.graphRD import convert_smile_to_mol_rdkit as rdkit_smile_to_mol
+    from kgcnn.mol.module_rdkit import convert_smile_to_mol_rdkit as rdkit_smile_to_mol
 except ImportError:
     module_logger.error("Can not import RDKit module for conversion.")
     rdkit_smile_to_mol = None
 
 try:
-    from kgcnn.mol.graphBabel import convert_smile_to_mol_openbabel as openbabel_smile_to_mol
+    from kgcnn.mol.module_babel import convert_smile_to_mol_openbabel as openbabel_smile_to_mol
 except ImportError:
     module_logger.error("Can not import OpenBabel module for conversion.")
     openbabel_smile_to_mol = None