+++ kgcnn/mol/module_pyxtal.py
@@ -1,0 +1,4 @@
+import pyxtal
+# pyxtal requires pymatgen as dependency.
+from pymatgen.core.structure import Structure
+