+++ kgcnn/utils/data.py
@@ -1,3 +1,5 @@
 from kgcnn.data.utils import *
 
 # For old reference.
+logging.error(
+    "This module is deprecated and will be removed in future versions. Please move to 'kgcnn.data.utils'.")