+++ kgcnn/utils/loss.py
@@ -1,7 +1,8 @@
 import tensorflow as tf
-
+import logging
 from kgcnn.metrics.metrics import *
 
 
 # For old reference. Moved to kgcnn.training module.
-
+logging.error(
+    "This module is deprecated and will be removed in future versions. Please move to 'kgcnn.metrics.metrics'.")